requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
flask==3.0.0
flask-cors==4.0.0
yfinance==0.2.33
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
transformers==4.36.0
tensorflow==2.15.0
python-dateutil==2.8.2
joblib==1.3.2
pydantic==2.5.3
python-dotenv==1.0.0
google-generativeai==0.3.2
aiohttp==3.9.3
protobuf==4.25.1
gunicorn
